requirements.txt
weaviate-client
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain[mistralai]
langchain[groq]
langchain-community
fastapi
langserve
uvicorn
sse_starlette
ipykernel
beautifulsoup4
pypdf
langchain-weaviate
wikipedia
arxiv
sentence-transformers